+++ ntex/src/http/h1/service.rs
@@ -233,12 +233,6 @@ where
     async fn shutdown(&self) {
         self.config.shutdown();
 
-        join(
-            self.config.control.shutdown(),
-            self.config.service.shutdown(),
-        )
-        .await;
-
         // check inflight connections
         let inflight = {
             let inflight = self.inflight.borrow();
@@ -256,6 +250,12 @@ where
 
             log::trace!("Shutting down is complected",);
         }
+
+        join(
+            self.config.control.shutdown(),
+            self.config.service.shutdown(),
+        )
+        .await;
     }
 
     async fn call(&self, io: Io<F>, _: ServiceCtx<'_, Self>) -> Result<(), Self::Error> {