+++ ntex/src/http/h2/service.rs
@@ -1,7 +1,9 @@
-use std::{cell::RefCell, error::Error, fmt, future::poll_fn, io, marker, mem, rc::Rc};
+use std::cell::{Cell, RefCell};
+use std::{error::Error, fmt, future::poll_fn, io, marker, mem, rc::Rc};
 
 use ntex_h2::{self as h2, frame::StreamId, server};
 
+use crate::channel::oneshot;
 use crate::http::body::{BodySize, MessageBody};
 use crate::http::config::{DispatcherConfig, ServiceConfig};
 use crate::http::error::{DispatchError, H2Error, ResponseError};
@@ -10,7 +12,7 @@ use crate::http::message::{CurrentIo, ResponseHead};
 use crate::http::{DateService, Method, Request, Response, StatusCode, Uri, Version};
 use crate::io::{types, Filter, Io, IoBoxed, IoRef};
 use crate::service::{IntoServiceFactory, Service, ServiceCtx, ServiceFactory};
-use crate::util::{Bytes, BytesMut, HashMap};
+use crate::util::{Bytes, BytesMut, HashMap, HashSet};
 
 use super::payload::{Payload, PayloadSender};
 use super::DefaultControlService;
@@ -177,11 +179,16 @@ where
             .create(())
             .await
             .map_err(|e| log::error!("Cannot construct publish service: {:?}", e))?;
+
+        let (tx, rx) = oneshot::channel();
         let config = Rc::new(DispatcherConfig::new(self.cfg.clone(), service, ()));
 
         Ok(H2ServiceHandler {
             config,
             control: self.ctl.clone(),
+            inflight: RefCell::new(Default::default()),
+            rx: Cell::new(Some(rx)),
+            tx: Cell::new(Some(tx)),
             _t: marker::PhantomData,
         })
     }
@@ -191,6 +198,9 @@ where
 pub struct H2ServiceHandler<F, S: Service<Request>, B, C> {
     config: Rc<DispatcherConfig<S, ()>>,
     control: Rc<C>,
+    inflight: RefCell<HashSet<IoRef>>,
+    rx: Cell<Option<oneshot::Receiver<()>>>,
+    tx: Cell<Option<oneshot::Sender<()>>>,
     _t: marker::PhantomData<(F, B)>,
 }
 
@@ -218,6 +228,25 @@ where
 
     #[inline]
     async fn shutdown(&self) {
+        self.config.shutdown();
+
+        // check inflight connections
+        let inflight = {
+            let inflight = self.inflight.borrow();
+            for io in inflight.iter() {
+                io.notify_dispatcher();
+            }
+            inflight.len()
+        };
+        if inflight != 0 {
+            log::trace!("Shutting down service, in-flight connections: {}", inflight);
+            if let Some(rx) = self.rx.take() {
+                let _ = rx.await;
+            }
+
+            log::trace!("Shutting down is complected",);
+        }
+
         self.config.service.shutdown().await
     }
 
@@ -226,9 +255,16 @@ where
         io: Io<F>,
         _: ServiceCtx<'_, Self>,
     ) -> Result<Self::Response, Self::Error> {
+        let inflight = {
+            let mut inflight = self.inflight.borrow_mut();
+            inflight.insert(io.get_ref());
+            inflight.len()
+        };
+
         log::trace!(
-            "New http2 connection, peer address {:?}",
-            io.query::<types::PeerAddr>().get()
+            "New http2 connection, peer address {:?}, inflight: {}",
+            io.query::<types::PeerAddr>().get(),
+            inflight
         );
         let control = self.control.create(()).await.map_err(|e| {
             DispatchError::Control(
@@ -236,7 +272,20 @@ where
             )
         })?;
 
-        handle(io.into(), control, self.config.clone()).await
+        let ioref = io.get_ref();
+        let result = handle(io.into(), control, self.config.clone()).await;
+        {
+            let mut inflight = self.inflight.borrow_mut();
+            inflight.remove(&ioref);
+
+            if inflight.len() == 0 {
+                if let Some(tx) = self.tx.take() {
+                    let _ = tx.send(());
+                }
+            }
+        }
+
+        result
     }
 }